+++ src/main/java/pl/indexpz/iStat/domain/service/impl/VehicleServiceImpl.java
@@ -10,25 +10,23 @@ import pl.indexpz.iStat.domain.model.Vehicle;
 import pl.indexpz.iStat.domain.repository.UserRepository;
 import pl.indexpz.iStat.domain.repository.VehicleRepository;
 import pl.indexpz.iStat.domain.service.VehicleService;
-import pl.indexpz.iStat.exceptions.ResourceNotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor //zamiast konstruktora
 public class VehicleServiceImpl implements VehicleService {
 
-    private final VehicleRepository vehicleRepository;
     private final UserRepository userRepository;
-
+    private final VehicleRepository vehicleRepository;
 
     @Override
     @Transactional
     public Vehicle addVehicle(Vehicle vehicleToAdd) {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByUsername(username).get();
-
         vehicleToAdd.setUser(user);
         return vehicleRepository.save(vehicleToAdd);
     }
@@ -40,8 +38,8 @@ public class VehicleServiceImpl implements VehicleService {
     }
 
     @Override
-    public Vehicle getVehicleById(Long id) {
-        return vehicleRepository.findById(id).orElseThrow((() -> new ResourceNotFoundException("Vehicle with id " + id + " not exist.")));
+    public Optional<Vehicle> getVehicleById(Long id) {
+        return vehicleRepository.findById(id);
     }
 
     @Override
@@ -50,20 +48,30 @@ public class VehicleServiceImpl implements VehicleService {
         return vehicle;
     }
 
-
     @Override
     public void updateVehicle(Vehicle vehicleToUpdate) {
-        Vehicle vehicle = getVehicleById(vehicleToUpdate.getId());
-        vehicle.setVehicleName(vehicleToUpdate.getVehicleName());
-        vehicle.setFuelName(vehicleToUpdate.getFuelName());
-        vehicle.setFuelUnit(vehicle.getFuelUnit());
-        vehicle.setMeterUnit(vehicleToUpdate.getMeterUnit());
-        vehicleRepository.save(vehicle);
+        Optional<Vehicle>  optionalVehicle = getVehicleById(vehicleToUpdate.getId());
+        if(optionalVehicle.isPresent()){
+            Vehicle vehicle = optionalVehicle.get();
+            vehicle.setVehicleName(vehicleToUpdate.getVehicleName());
+            vehicle.setFuelName(vehicleToUpdate.getFuelName());
+            vehicle.setFuelUnit(vehicle.getFuelUnit());
+            vehicle.setMeterUnit(vehicleToUpdate.getMeterUnit());
+            vehicleRepository.save(vehicle);
+        }else{
+            Vehicle vehicle = new Vehicle();
+        }
     }
 
     @Override
     public void removeVehicle(Vehicle vehicleToDelete) {
-        Vehicle vehicle = getVehicleById(vehicleToDelete.getId());
-        vehicleRepository.delete(vehicle);
+        Optional<Vehicle>  optionalVehicle = getVehicleById(vehicleToDelete.getId());
+        if(optionalVehicle.isPresent()){
+            Vehicle vehicle = optionalVehicle.get();
+            vehicleRepository.delete(vehicle);
+        }else{
+            Vehicle vehicle = new Vehicle();
+        }
+
     }
 }
